Function massa_protocol_worker::handlers::endorsement_handler::retrieval::note_endorsements_from_peer

pub(crate) fn note_endorsements_from_peer(
    endorsements: Vec<SecureShareEndorsement>,
    from_peer_id: &PeerId,
    cache: &Arc<RwLock<EndorsementCache>>,
    selector_controller: &dyn SelectorController,
    storage: &Storage,
    config: &ProtocolConfig,
    endorsement_propagation_sender: &MassaSender<EndorsementHandlerPropagationCommand>,
    pool_controller: &mut dyn PoolController,
) -> Result<(), ProtocolError>

Note endorsements coming from a given node, and propagate them when they were received outside of a header.

Caches knowledge of valid ones.

Does not ban if the endorsement is invalid

Checks performed:

Valid signature.