Struct massa_protocol_worker::protocol_worker::ProtocolWorker

pub struct ProtocolWorker {

    pub(crate) config: ProtocolConfig,
    pub(crate) consensus_controller: Box<dyn ConsensusController>,
    pub(crate) network_command_sender: NetworkCommandSender,
    network_event_receiver: NetworkEventReceiver,
    pool_controller: Box<dyn PoolController>,
    controller_command_rx: Receiver<ProtocolCommand>,
    controller_manager_rx: Receiver<ProtocolManagementCommand>,
    pub(crate) active_nodes: HashMap<NodeId, NodeInfo>,
    pub(crate) block_wishlist: PreHashMap<BlockId, BlockInfo>,
    checked_endorsements: LinearHashCacheSet<EndorsementId>,
    pub(crate) checked_operations: CheckedOperations,
    pub(crate) checked_headers: LinearHashCacheMap<BlockId, SecuredHeader>,
    pub(crate) asked_operations: PreHashMap<OperationPrefixId, (Instant, Vec<NodeId>)>,
    pub(crate) op_batch_buffer: OperationBatchBuffer,
    pub(crate) storage: Storage,
    operations_to_announce: Vec<OperationId>,

}

protocol worker

Fields
config: ProtocolConfig

Protocol configuration.

consensus_controller: Box<dyn ConsensusController>

Consensus controller

network_command_sender: NetworkCommandSender

Associated network command sender.

network_event_receiver: NetworkEventReceiver

Associated network event receiver.

pool_controller: Box<dyn PoolController>

Channel to send protocol pool events to the controller.

controller_command_rx: Receiver<ProtocolCommand>

Channel receiving commands from the controller.

controller_manager_rx: Receiver<ProtocolManagementCommand>

Channel to send management commands to the controller.

active_nodes: HashMap<NodeId, NodeInfo>

Ids of active nodes mapped to node info.

block_wishlist: PreHashMap<BlockId, BlockInfo>

List of wanted blocks, with the info representing their state with in the as_block workflow.

checked_endorsements: LinearHashCacheSet<EndorsementId>

List of processed endorsements

checked_operations: CheckedOperations

Cache of processed operations

checked_headers: LinearHashCacheMap<BlockId, SecuredHeader>

List of processed headers

asked_operations: PreHashMap<OperationPrefixId, (Instant, Vec<NodeId>)>

List of ids of operations that we asked to the nodes

op_batch_buffer: OperationBatchBuffer

Buffer for operations that we want later

storage: Storage

Shared storage.

operations_to_announce: Vec<OperationId>

Operations to announce at the next interval.

Implementations

impl ProtocolWorker

pub fn new( config: ProtocolConfig, _: ProtocolWorkerChannels, consensus_controller: Box<dyn ConsensusController>, pool_controller: Box<dyn PoolController>, storage: Storage ) -> ProtocolWorker

Creates a new protocol worker.

Arguments

config: protocol configuration.

network_controller: associated network controller.

controller_event_tx: Channel to send protocol events.

controller_command_rx: Channel receiving commands.

controller_manager_rx: Channel receiving management commands.

pub async fn run_loop(self) -> Result<NetworkEventReceiver, ProtocolError>

Main protocol worker loop. Consumes self. It is mostly a tokio::select! inside a loop waiting on :

controller_command_rx

network_controller

handshake_futures

node_event_rx And at the end every thing is closed properly Consensus work is managed here. It’s mostly a tokio::select! within a loop.

async fn announce_ops(&mut self, timer: &mut Pin<&mut Sleep>)

Announce a set of operations to active nodes who do not know about it yet. Side effects:

notes nodes as knowing about those operations from now on.

empties the buffer of operations to announce.

async fn note_operations_to_announce( &mut self, operations: &[OperationId], timer: &mut Pin<&mut Sleep> )

Add an list of operations to a buffer for announcement at the next interval, or immediately if the buffer is full.

async fn propagate_endorsements(&mut self, storage: &Storage)

async fn process_command( &mut self, cmd: ProtocolCommand, block_timer: &mut Pin<&mut Sleep>, op_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

pub(crate) fn remove_asked_blocks_of_node( &mut self, remove_hashes: &PreHashSet<BlockId> ) -> Result<(), ProtocolError>

Remove the given blocks from the local wishlist

pub(crate) async fn update_ask_block( &mut self, ask_block_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

pub(crate) async fn ban_node( &mut self, node_id: &NodeId ) -> Result<(), ProtocolError>

Ban a node.

pub(crate) async fn note_header_from_node( &mut self, header: &SecuredHeader, source_node_id: &NodeId ) -> Result<Option<(BlockId, bool)>, ProtocolError>

Perform checks on a header, and if valid update the node’s view of the world.

Returns a boolean representing whether the header is new.

Does not ban the source node if the header is invalid.

Checks performed on Header:

Not genesis.

Can compute a BlockId.

Valid signature.

Absence of duplicate endorsements.

Checks performed on endorsements:

Unique indices.

Slot matches that of the block.

Block matches that of the block.

pub(crate) async fn note_operations_from_node( &mut self, operations: Vec<SecureShareOperation>, source_node_id: &NodeId, op_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

Checks operations, caching knowledge of valid ones.

Does not ban if the operation is invalid.

Checks performed:

Valid signature

pub(crate) async fn note_endorsements_from_node( &mut self, endorsements: Vec<SecureShareEndorsement>, source_node_id: &NodeId, propagate: bool ) -> Result<(), ProtocolError>

Note endorsements coming from a given node, and propagate them when they were received outside of a header.

Caches knowledge of valid ones.

Does not ban if the endorsement is invalid

Checks performed:

Valid signature.

impl ProtocolWorker

pub(crate) async fn on_operations_announcements_received( &mut self, op_batch: OperationPrefixIds, node_id: NodeId ) -> Result<(), ProtocolError>

On receive a batch of operation ids op_batch from another node_id Execute the following algorithm: redirect to GitHub

def process_op_batch(op_batch, node_id):
    ask_set = void HashSet<OperationId>
    future_set = void HashSet<OperationId>
    for op_id in op_batch:
        if not is_op_received(op_id):
            if (op_id not in asked_ops) or (node_id not in asked_ops(op_id)[1]):
                if (op_id not in asked_ops) or (asked_ops(op_id)[0] < now - op_batch_proc_period:
                    ask_set.add(op_id)
                    asked_ops(op_id)[0] = now
                    asked_ops(op_id)[1].add(node_id)
                else:
                    future_set.add(op_id)
    if op_batch_buf is not full:
        op_batch_buf.push(now+op_batch_proc_period, node_id, future_set)
    ask ask_set to node_id

pub(crate) async fn on_operations_received( &mut self, node_id: NodeId, operations: Vec<SecureShareOperation>, op_timer: &mut Pin<&mut Sleep> )

On full operations are received from the network,

Update the cache received_operations ids and each node_info.known_operations

Notify the operations to he local node, to be propagated

pub(crate) fn prune_asked_operations( &mut self, ask_operations_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

Clear the asked_operations data structure and reset ask_operations_timer

pub(crate) async fn update_ask_operation( &mut self, operation_batch_proc_period_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

pub(crate) async fn on_asked_operations_received( &mut self, node_id: NodeId, op_pre_ids: OperationPrefixIds ) -> Result<(), ProtocolError>

Process the reception of a batch of asked operations, that means that we have already sent a batch of ids in the network, notifying that we already have those operations.

impl ProtocolWorker

pub(crate) async fn on_network_event( &mut self, evt: NetworkEvent, block_ask_timer: &mut Pin<&mut Sleep>, op_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

Manages network event Only used by the worker.

Argument

evt: event to process block_ask_timer: Timer to update to the next time we are able to ask a block

async fn on_asked_for_blocks_received( &mut self, from_node_id: NodeId, list: Vec<(BlockId, AskForBlocksInfo)> ) -> Result<(), ProtocolError>

Network ask the local node for blocks

React on another node asking for blocks information. We can forward the operation ids if the foreign node asked for AskForBlocksInfo::Info or the full operations if he asked for the missing operations in his storage with AskForBlocksInfo::Operations

Forward the reply to the network.

fn get_total_operations_size( storage: &Storage, operation_ids: &PreHashSet<OperationId> ) -> usize

Return the sum of all operation’s serialized sizes in the Set<Id>

async fn on_block_header_received( &mut self, from_node_id: NodeId, block_id: BlockId, header: SecuredHeader ) -> Result<(), ProtocolError>

On block header received from a node. If the header is new, we propagate it to the consensus. We pass the state of block_wishlist to ask for information about the block.

async fn on_block_operation_list_received( &mut self, from_node_id: NodeId, block_id: BlockId, operation_ids: Vec<OperationId>, op_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

On block information received, manage when we get a list of operations. Ask for the missing operations that are not in the checked_operations cache variable.

Ban

Start compute the operations serialized total size with the operation we know. Ban the node if the operations contained in the block overflow the max size. We don’t forward the block to the consensus in that case.

Parameters:

from_node_id: Node which sent us the information.

BlockId: ID of the related operations we received.

operation_ids: IDs of the operations contained by the block.

Result

return an error if stopping asking block failed. The error should be forwarded at the root. todo: check if if make panic.

async fn on_block_full_operations_received( &mut self, from_node_id: NodeId, block_id: BlockId, operations: Vec<SecureShareOperation>, op_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>

Checks full block operations that we asked. (Because their was missing in the checked_operations cache variable, refer to on_block_operation_list_received)

Ban

Ban the node if it doesn’t fill the requirement. Forward to the graph with a ProtocolEvent::ReceivedBlock if the operations are under a max size.

thread incorrect for an operation

wanted operations doesn’t match

duplicated operation

full operations serialized size overflow

We received these operation because we asked for the missing operation

async fn on_block_info_received( &mut self, from_node_id: NodeId, block_id: BlockId, info: BlockInfoReply, op_timer: &mut Pin<&mut Sleep> ) -> Result<(), ProtocolError>