Function massa_protocol_worker::worker::start_protocol_controller

pub fn start_protocol_controller(
    config: ProtocolConfig,
    selector_controller: Box<dyn SelectorController>,
    consensus_controller: Box<dyn ConsensusController>,
    bootstrap_peers: Option<BootstrapPeers>,
    pool_controller: Box<dyn PoolController>,
    storage: Storage,
    protocol_channels: ProtocolChannels,
    mip_store: MipStore,
    massa_metrics: MassaMetrics
) -> Result<(Box<dyn ProtocolManager>, KeyPair, NodeId), ProtocolError>

start a new ProtocolController from a ProtocolConfig

Arguments

config: protocol settings

consensus_controller: interact with consensus module

bootstrap_peers: list of peers to connect to retrieved from the bootstrap

storage: Shared storage to fetch data that are fetch across all modules