Module massa_protocol_worker::worker_operations_impl

Contains the implementation of the life cycle of operations

Implement the propagation algorithm written here [redirect to GitHub] (https://github.com/massalabs/massa/issues/2283#issuecomment-1040872779).

get batches of operations ids

ask for operations

send batches

answer operations

Structs

OperationBatchItem

Structure containing a Batch of operation_ids we would like to ask to a node_id now or later. Mainly used in protocol and translated into simple combination of a node_id and operations_prefix_ids

Type Definitions

OperationBatchBuffer

Queue containing every [OperationsBatchItem] we want to ask now or later.