Crate massa_protocol_worker

Protocol component High level management of communications between nodes

Re-exports

pub use protocol_worker::start_protocol_controller;

Modules

cache 🔒

Simple unreliable, but fast cache implementations

checked_operations 🔒

Cache of previously successfully checked operations and their prefix IDs

node_info 🔒

Only public for the protocol worker. Contains all what we know for each node we are connected or we used to know in the network.

protocol_network 🔒

Worker implementation for the network events

protocol_worker

protocol worker Copyright (c) 2022 MASSA LABS info@massa.net

sig_verifier 🔒

Optimized batch signature verifier

worker_operations_impl

Contains the implementation of the life cycle of operations